// src/lib/server/schemas/rsstream.ts
export const updatePasswordSchema = z.object({
  code: z.string().min(1),
  reset: z.string().min(1),
  pub: z.string().min(1),
  epub: z.string().min(1),
  username: z.string().min(1),
  name: z.string().min(1),
})